requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
werkzeug==3.0.1
pandas
openpyxl
PyPDF2
python-docx
scikit-learn
numpy
